+++ src/main/java/dgu/edu/dnaapi/domain/dto/auth/RefreshTokenDto.java
@@ -1,4 +1,4 @@
-package dgu.edu.dnaapi.domain.dto;
+package dgu.edu.dnaapi.domain.dto.auth;
 
 import lombok.Data;
 import lombok.NoArgsConstructor;
