+++ src/main/java/dgu/edu/dnaapi/domain/dto/auth/TokenResponse.java
@@ -1,4 +1,4 @@
-package dgu.edu.dnaapi.domain.dto;
+package dgu.edu.dnaapi.domain.dto.auth;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
