+++ src/main/java/dgu/edu/dnaapi/domain/dto/noticeComments/NoticeCommentsUpdateDto.java
@@ -1,4 +1,4 @@
-package dgu.edu.dnaapi.domain.dto;
+package dgu.edu.dnaapi.domain.dto.noticeComments;
 
 import lombok.Builder;
 import lombok.Getter;
