+++ src/main/java/dgu/edu/dnaapi/service/NoticesService.java
@@ -1,19 +1,15 @@
 package dgu.edu.dnaapi.service;
 
 import dgu.edu.dnaapi.domain.Notices;
-import dgu.edu.dnaapi.domain.User;
-import dgu.edu.dnaapi.domain.dto.NoticesDeleteRequestDto;
-import dgu.edu.dnaapi.domain.dto.NoticesResponseDto;
-import dgu.edu.dnaapi.domain.dto.NoticesSaveRequestDto;
-import dgu.edu.dnaapi.domain.dto.NoticesUpdateRequestDto;
+import dgu.edu.dnaapi.domain.dto.notices.NoticesDeleteRequestDto;
+import dgu.edu.dnaapi.domain.dto.notices.NoticesResponseDto;
+import dgu.edu.dnaapi.domain.dto.notices.NoticesUpdateRequestDto;
 import dgu.edu.dnaapi.repository.NoticesRepository;
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
-import java.util.Optional;
 import java.util.stream.Collectors;
 
 @RequiredArgsConstructor
